+++ engine/lib/membrane_rtc_engine/engine.ex
@@ -422,20 +422,15 @@ defmodule Membrane.RTC.Engine do
         :ignored
 
       {^ref, {:error, :endpoint_not_exist}} ->
-        Membrane.Logger.error(%{
+        Membrane.Logger.error("Couldn't subscribe to track",
           error: :endpoint_not_exist,
-          message: "Couldn't subscribe to track",
           track_id: track_id
-        })
+        )
 
         raise "Couldn't subscribe to the track: #{inspect(track_id)}, because endpoint #{endpoint_id} doesn't exist."
     after
       5_000 ->
-        Membrane.Logger.error(%{
-          error: :timeout,
-          message: "Couldn't subscribe to track",
-          track_id: track_id
-        })
+        Membrane.Logger.error("Couldn't subscribe to track", error: :timeout, track_id: track_id)
 
         raise "Couldn't subscribe to the track: #{inspect(track_id)}, because of timeout."
     end