+++ ex_webrtc/integration_test/test_videoroom/assets/src/types.ts
@@ -1,0 +1,6 @@
+export type EncodingStats = {
+  framesSent: number,
+  height: number,
+  width: number,
+  framesPerSecond: number
+}