+++ ex_webrtc/integration_test/test_videoroom/test/integration/basic_test.exs
@@ -56,7 +56,7 @@ defmodule TestVideoroom.Integration.BasicTest do
           {:after_warmup, browsers} ->
             Enum.each(browsers, fn {browser_id, stats_list} ->
               Enum.each(stats_list, fn stats ->
-                assert length(stats) == browser_id
+                assert length(stats) == 2 * browser_id
                 assert Enum.all?(stats, &is_stream_playing(&1))
               end)
             end)
@@ -64,7 +64,7 @@ defmodule TestVideoroom.Integration.BasicTest do
           {:before_leave, browsers} ->
             Enum.each(browsers, fn {browser_id, stats_list} ->
               Enum.each(stats_list, fn stats ->
-                assert length(stats) == browsers_number - browser_id - 1
+                assert length(stats) == 2 * (browsers_number - browser_id - 1)
                 assert Enum.all?(stats, &is_stream_playing(&1))
               end)
             end)
@@ -91,7 +91,6 @@ defmodule TestVideoroom.Integration.BasicTest do
 
     for browser <- 0..(browsers_number - 1), into: [] do
       mustang_options = %{mustang_options | id: browser}
-      Process.sleep(500)
 
       Task.async(fn ->
         Stampede.start({TestMustang, mustang_options}, @browser_options)
@@ -105,7 +104,7 @@ defmodule TestVideoroom.Integration.BasicTest do
           {:after_warmup, browsers} ->
             Enum.each(browsers, fn {_browser_id, stats_list} ->
               Enum.each(stats_list, fn stats ->
-                assert length(stats) == browsers_number - 1
+                assert length(stats) == 2 * (browsers_number - 1)
                 assert Enum.all?(stats, &is_stream_playing(&1))
               end)
             end)
@@ -149,7 +148,7 @@ defmodule TestVideoroom.Integration.BasicTest do
     end
     |> Task.await_many(:infinity)
 
-    {_button, buttons_with_id} = Map.pop!(buttons_with_id, 3)
+    browser_received_tracks = %{0 => 2, 1 => 3, 2 => 3, 3 => 4}
 
     receive do
       {:stats, acc} ->
@@ -157,10 +156,8 @@ defmodule TestVideoroom.Integration.BasicTest do
           {:after_warmup, browsers} ->
             Enum.each(browsers, fn {browser_id, stats_list} ->
               Enum.each(stats_list, fn stats ->
-                assert length(stats) == if(browser_id == 3, do: 3, else: 2)
-                {_value, new_buttons} = Map.pop(buttons_with_id, browser_id)
-                new_buttons = Map.values(new_buttons)
-                assert_streams_playing(stats, new_buttons)
+                assert length(stats) == browser_received_tracks[browser_id]
+                assert Enum.all?(stats, &is_stream_playing(&1))
               end)
             end)
 
@@ -170,29 +167,7 @@ defmodule TestVideoroom.Integration.BasicTest do
     end
   end
 
-  defp assert_streams_playing(stats, buttons) do
-    for button <- buttons do
-      case button do
-        @start_with_all ->
-          assert Enum.any?(stats, &is_stream_playing(&1))
-
-        @start_with_camera ->
-          assert Enum.any?(stats, &is_stream_playing(&1, %{audio: false, video: true}))
-
-        @start_with_mic ->
-          assert Enum.any?(stats, &is_stream_playing(&1, %{audio: true, video: false}))
-
-        @start_with_nothing ->
-          assert Enum.any?(stats, &is_stream_playing(&1, %{audio: false, video: false}))
-      end
-    end
+  defp is_stream_playing(%{"streamId" => _, "playing" => playing, "kind" => _kind}) do
+    playing == true
   end
-
-  defp is_stream_playing(stats, expected \\ %{audio: true, video: true})
-
-  defp is_stream_playing(
-         %{"streamId" => _, "isAudioPlaying" => audio, "isVideoPlaying" => video},
-         %{audio: expected_audio, video: expected_video}
-       ),
-       do: audio == expected_audio and video == expected_video
 end