+++ ex_webrtc/integration_test/test_videoroom/test/integration/simulcast_test_off.exs
@@ -1,6 +1,5 @@
 defmodule TestVideoroom.Integration.SimulcastTest do
   use TestVideoroomWeb.ConnCase, async: false
-
   import TestVideoroom.Integration.Utils
 
   @room_url "http://localhost:4001"