+++ engine/CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add reason to `TrackVariantPaused` event [#392](https://github.com/jellyfish-dev/membrane_rtc_engine/pull/392)
 * Lower log level for some logs [#401](https://github.com/fishjam-dev/membrane_rtc_engine/pull/401)
 * Add async subscribe [#407](https://github.com/fishjam-dev/membrane_rtc_engine/pull/407)
+* Add ex_webrtc integration test [#409](https://github.com/fishjam-dev/membrane_rtc_engine/pull/409)
 
 ## 0.22.0
 * Update deps [#374](https://github.com/jellyfish-dev/membrane_rtc_engine/pull/374)